Stock Tracking System
StockTrackingSystem is a web-based application designed to track and analyze live stock market data. The system fetches real-time stock information, processes it, and displays comprehensive details such as prices, volume, and percentage changes. Additionally, it integrates with an MS SQL database to store and manage stock data for further analysis.